+++ fairmq/Tools.h
@@ -1,0 +1,17 @@
+/********************************************************************************
+ *    Copyright (C) 2017 GSI Helmholtzzentrum fuer Schwerionenforschung GmbH    *
+ *                                                                              *
+ *              This software is distributed under the terms of the             *
+ *              GNU Lesser General Public Licence (LGPL) version 3,             *
+ *                  copied verbatim in the file "LICENSE"                       *
+ ********************************************************************************/
+
+#ifndef FAIR_MQ_TOOLS_H
+#define FAIR_MQ_TOOLS_H
+
+// IWYU pragma: begin_exports
+#include <fairmq/tools/CppSTL.h>
+#include <fairmq/tools/Network.h>
+// IWYU pragma: end_exports
+
+#endif // FAIR_MQ_TOOLS_H
